refactor: :recycle: Separated API call to service.ts
